# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateDailyScheduleRequest:
    properties:
      company_id:
        type: string
      date:
        type: string
      employee_id:
        type: string
      end_time:
        type: string
      remarks:
        type: string
      schedule_type:
        type: string
      start_time:
        type: string
    required:
    - company_id
    - date
    - employee_id
    - schedule_type
    type: object
...
  service.LoginRequest:
    properties:
      email:
        type: string
      mfa_code:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  service.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/service.UserResponse'
    type: object
  service.ProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      force_password_change:
        type: boolean
      id:
        type: string
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/service.RoleInfo'
        type: array
      status:
        type: string
    type: object
...
  service.RoleInfo:
    properties:
      name:
        type: string
    type: object
  service.UpdateProfileRequest:
    properties:
      name:
        type: string
    type: object
...
  service.UserResponse:
    properties:
      email:
        type: string
      force_password_change:
        type: boolean
      id:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
host: localhost:5000
info:
  contact:
    email: support@peoplehub.com
    name: PeopleHub Team
  description: Enterprise HR Management & Payroll System API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PeopleHub API
  version: "1.1"
paths:
...
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Change user password after verifying current password
      parameters:
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - Auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revoke refresh token to logout
      parameters:
      - description: Refresh token to revoke
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - Auth
  /auth/logout-all:
    post:
      description: Revoke all refresh tokens for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout all sessions
      tags:
      - Auth
  /auth/me:
    get:
      description: Get current authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Auth
  /auth/profile:
    put:
      consumes:
      - application/json
      description: Update current authenticated user's profile (name)
      parameters:
      - description: Profile update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange refresh token for new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - Auth
...
  /auth/sessions:
    get:
      description: List all active sessions for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active sessions
      tags:
      - Auth
...
  /daily-schedules:
    get:
      description: Get all daily schedules for a company, optionally filtered by date
      parameters:
      - description: Company ID
        in: query
        name: company_id
        type: string
      - description: Date (YYYY-MM-DD)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List daily schedules
      tags:
      - HR
    post:
      consumes:
      - application/json
      description: Create a new daily schedule for an employee
      parameters:
      - description: Schedule details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateDailyScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create daily schedule
      tags:
      - HR
  /daily-schedules/{id}:
    delete:
      description: Soft delete a daily schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete daily schedule
      tags:
      - HR
    put:
      consumes:
      - application/json
      description: Update an existing daily schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.CreateDailyScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update daily schedule
      tags:
      - HR
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"